+++ src/templates/docker/docker-code-server/main.tf
@@ -101,47 +101,6 @@ data "coder_parameter" "repo" {
   order       = 2
 }
 
-data "coder_parameter" "git_user_name" {
-  type        = "string"
-  name        = "Git user.name"
-  description = "Used to run: git config --global user.name"
-  default     = ""
-  mutable     = true
-  icon        = "/emojis/1f511.png"
-  order       = 3 
-}
-
-data "coder_parameter" "git_user_email" {
-  type        = "string"
-  name        = "Git user.email"
-  description = "Used to run: git config --global user.email"
-  default     = ""
-  mutable     = true
-  icon        = "/emojis/1f511.png"
-  order       = 4  
-}
-
-
-data "coder_parameter" "github_user_name" {
-  type        = "string"
-  name        = "GitHub username"
-  description = "Used to run: git config --global credential...username"
-  default     = ""
-  mutable     = true
-  icon        = "/emojis/1f511.png"
-  order       = 5  
-}
-
-data "coder_parameter" "github_personal_access_token" {
-  type        = "string"
-  name        = "GitHub personal access token"
-  description = "Used to run with git credential-store store"
-  default     = ""
-  mutable     = true
-  icon        = "/emojis/1f511.png"
-  order       = 6
-}
-
 module "dotfiles" {
   count    = data.coder_workspace.me.start_count
   source   = "registry.coder.com/modules/dotfiles/coder"
@@ -168,6 +127,12 @@ module "git-clone" {
 module "coder-login" {
   count    = data.coder_workspace.me.start_count
   source   = "registry.coder.com/modules/coder-login/coder"
+  agent_id = coder_agent.dev.id
+}
+
+module "git-config" {
+  count    = data.coder_workspace.me.start_count
+  source   = "registry.coder.com/modules/git-config/coder"
   agent_id = coder_agent.dev.id
 }
 
@@ -234,29 +199,19 @@ resource "coder_agent" "dev" {
   startup_script  = <<EOT
   #!/bin/sh
 
-  # configure git username and email for commits
-  if [ ! -z "${data.coder_parameter.git_user_name.value}" ]; then
-    git config --global user.name "${data.coder_parameter.git_user_name.value}"
-  fi
-  if [ ! -z "${data.coder_parameter.git_user_email.value}" ]; then
-    git config --global user.email "${data.coder_parameter.git_user_email.value}"
-  fi
-  # configure git credential helper for github
-  if [ ! -z "${data.coder_parameter.github_user_name.value}" ]; then
-    git config --global credential.https://github.com.username "${data.coder_parameter.github_user_name.value}"
-  fi
-
-  if [ ! -z "${data.coder_parameter.github_personal_access_token.value}" ]; then
-    git config --global credential.helper store
-    printf "protocol=https\nhost=github.com\nusername=%s\npassword=%s\n" "${data.coder_parameter.github_user_name.value}" "${data.coder_parameter.github_personal_access_token.value}" | git credential-store store
-  fi
+    if ! command -v aider >/dev/null 2>&1; then
+      curl -LsSf https://aider.chat/install.sh | sh
+    fi
+    if ! command -v goose >/dev/null 2>&1; then
+      curl -fsSL https://github.com/block/goose/releases/download/stable/download_cli.sh | CONFIGURE=false bash
+    fi
 
   EOT
 }
 
 resource "docker_container" "workspace" {
   count = data.coder_workspace.me.start_count
-  image = "${data.coder_parameter.image.value}"
+  image = data.coder_parameter.image.value
   # Uses lower() to avoid Docker restriction on container names.
   name     = "coder-${data.coder_workspace_owner.me.name}-${lower(data.coder_workspace.me.name)}"
   hostname = lower(data.coder_workspace.me.name)
@@ -285,6 +240,7 @@ resource "docker_container" "workspace" {
     host = "host.docker.internal"
     ip   = "host-gateway"
   }
+
 }
 
 resource "docker_volume" "coder_volume" {